just deleting comments and rearranging files
