+++ cmd_cc_add.go
@@ -3,10 +3,8 @@ package main
 import (
 	"bufio"
 	"context"
-	"errors"
 	"fmt"
 	"os"
-	"strconv"
 	"strings"
 
 	"github.com/jsandberg07/clitest/internal/database"
@@ -25,8 +23,6 @@ func getAddCCCmd() Command {
 	return addCCCmd
 }
 
-// [s]tart, [e]nd, [a]dd, [i]nvestigator, [p]rotocol, help, exit, save
-// have to handle dupes as the input int is
 func getAddCCFlags() map[string]Flag {
 	addCCFlags := make(map[string]Flag)
 	sFlag := Flag{
@@ -69,8 +65,6 @@ func getAddCCFlags() map[string]Flag {
 	}
 	addCCFlags[pFlag.symbol] = pFlag
 
-	// ect as needed or remove the "-"+ for longer ones
-
 	helpFlag := Flag{
 		symbol:      "help",
 		description: "Prints available flags",
@@ -99,9 +93,7 @@ func getAddCCFlags() map[string]Flag {
 
 }
 
-// look into removing the args thing, might have to stay
 func addCCFunction(cfg *Config) error {
-	// permission check
 	err := checkPermission(cfg.loggedInPosition, PermissionActivateInactivate)
 	if err != nil {
 		return err
@@ -121,21 +113,16 @@ func addCCFunction(cfg *Config) error {
 		return err
 	}
 
-	// get flags
 	flags := getAddCCFlags()
 
-	// set defaults
 	exit := false
 	var start int
 	var end int
 
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
 
 	fmt.Println("Adding cards under logged in user. Change using flags.")
 	fmt.Println("Use the flags to set the range of cards to be added to the DB")
-	// fmt.Println("Enter help to see available flags")
-	// da loop
 	for {
 		fmt.Print("> ")
 		text, err := reader.ReadString('\n')
@@ -150,16 +137,12 @@ func addCCFunction(cfg *Config) error {
 			continue
 		}
 
-		// do weird behavior here
-
-		// but normal loop now
 		args, err := parseArguments(flags, inputs)
 		if err != nil {
 			fmt.Println(err)
 			continue
 		}
 
-		// [s]tart, [e]nd, [a]dd, [i]nvestigator, [p]rotocol, help, exit, save
 		for _, arg := range args {
 			switch arg.flag {
 			case "-s":
@@ -188,6 +171,10 @@ func addCCFunction(cfg *Config) error {
 				inv, err := getInvestigatorByFlag2(cfg, arg.value)
 				if err != nil {
 					return err
+				}
+				nilInvestigator := database.Investigator{}
+				if inv == nilInvestigator {
+					break
 				}
 
 				err = investigatorProtocolCheck(cfg, &inv, &protocol)
@@ -243,28 +230,12 @@ func addCCFunction(cfg *Config) error {
 	return nil
 }
 
-// kind of a place holder. TODO: decide what to do with error
-func getNumberFromFlag(input string) (int, error) {
-	if input == "" {
-		fmt.Println("No input found. Please try again.")
-		return 0, nil
-	}
-	num, err := strconv.Atoi(input)
-	if err != nil {
-		fmt.Println("Could not read number from input")
-		return 0, err
-	}
-	return num, nil
-}
-
-// possibly go routine this later, for now should be fast enough
 func addCCtoDB(cfg *Config, start, end int, inv database.Investigator, pro database.Protocol) error {
 	if end < start {
 		fmt.Println("Start of range is larger than end. Please check inputs and try again.")
 		return nil
 	}
 
-	// maybe remove
 	fmt.Printf("Adding %v cards\n", (end-start)+1)
 	added := 0
 	// inclusive. Not interating an array. The <= is intended.
@@ -292,40 +263,4 @@ func addCCtoDB(cfg *Config, start, end int, inv database.Investigator, pro datab
 	}
 	fmt.Printf("%v cards added!\n", added)
 	return nil
-}
-
-// literally rewrote this by accident so here's a smaller version, will decide on what to use later
-func getInvestigatorByFlag2(cfg *Config, input string) (database.Investigator, error) {
-	if input == "" {
-		fmt.Println("No input found, please try again")
-		return database.Investigator{}, nil
-	}
-	investigators, err := cfg.db.GetInvestigatorByName(context.Background(), input)
-	if err != nil {
-		fmt.Println("Error getting investigator from db")
-		return database.Investigator{}, err
-	}
-	if len(investigators) == 0 {
-		fmt.Println("No investigator by that name found. Nicknames also work as well.")
-		return database.Investigator{}, nil
-	}
-	if len(investigators) > 1 {
-		fmt.Println("Vague investigator name. Please try again.")
-		return database.Investigator{}, nil
-	}
-	return investigators[0], nil
-}
-
-func getProtocolStruct(cfg *Config, input string) (database.Protocol, error) {
-	protocol, err := cfg.db.GetProtocolByNumber(context.Background(), input)
-	if err != nil && err.Error() == "sql: no rows in result set" {
-		return database.Protocol{}, errors.New("protocol not found. please try again")
-	}
-	if err != nil && err.Error() != "sql: no rows in result set" {
-		// any other error
-		fmt.Println("Error getting strain from DB.")
-		return database.Protocol{}, err
-	}
-
-	return protocol, nil
 }
