+++ cmd_cc_deactivate.go
@@ -28,14 +28,6 @@ func getCCDeactivationCmd() Command {
 	return deactivationCmd
 }
 
-// enter number
-// try to deactivate
-// get error if need be
-// delete reminder
-// its easier trust me
-
-// cc, pop, process, exit, list of errors (both previously deact and not activated)
-// TODO: change this to work like activation (ie linear, check reminders)
 func getDeactivationFlags() map[string]Flag {
 	deactivationFlags := make(map[string]Flag)
 	ccFlag := Flag{
@@ -83,23 +75,17 @@ func getDeactivationFlags() map[string]Flag {
 }
 
 func deactivateFunction(cfg *Config) error {
-	// permission check
 	err := checkPermission(cfg.loggedInPosition, PermissionDeactivateReactivate)
 	if err != nil {
 		return err
 	}
-	// get flags
 	flags := getDeactivationFlags()
 
-	// set defaults
 	exit := false
 	date := normalizeDate(time.Now())
 
 	fmt.Println("Enter cards to deactivate.")
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
-
-	// da loop
 	for {
 		fmt.Print("> ")
 		text, err := reader.ReadString('\n')
@@ -114,9 +100,6 @@ func deactivateFunction(cfg *Config) error {
 			continue
 		}
 
-		// do weird behavior here
-
-		// try to run as a number, and add it to the list of cards to activate using the current values
 		if len(inputs) == 1 {
 			cc, err := strconv.Atoi(inputs[0])
 			if err != nil && !strings.Contains(err.Error(), "invalid syntax") {
@@ -125,28 +108,23 @@ func deactivateFunction(cfg *Config) error {
 				fmt.Println(err)
 				continue
 			}
-
 			// a misread on cc means the value 0 init
 			if cc != 0 {
 				err := deactivateCC(cfg, cc, date)
 				if err != nil {
 					fmt.Println(err)
 				}
-
 				// don't need to visit the switch, one input is assumed to be a cc#
 				continue
 			}
 		}
 
-		// but normal loop now
 		args, err := parseArguments(flags, inputs)
 		if err != nil {
 			fmt.Println(err)
 			continue
 		}
 
-		// cc, pop, process, exit, list of errors (both previously deact and not activated)
-		// TODO: make sure activated cards check if previously deact too
 		for _, arg := range args {
 			switch arg.flag {
 
@@ -157,8 +135,7 @@ func deactivateFunction(cfg *Config) error {
 					break
 				}
 
-				// cant be after today ie cant deactivate on a day that hasnt happened
-				if newDate.After(time.Now()) {
+				if newDate.After(normalizeDate(time.Now())) {
 					fmt.Println("Deactivation date can't be set in the future")
 					break
 				}
@@ -166,22 +143,22 @@ func deactivateFunction(cfg *Config) error {
 				date = normalizeDate(newDate)
 				fmt.Printf("Date set: %v\n", date)
 
-			// TODO: -cc isnt working for some reason but im not testing it atm
 			case "-cc":
-				cc, err := strconv.Atoi(inputs[0])
+				cc, err := strconv.Atoi(arg.value)
 				if err != nil && !strings.Contains(err.Error(), "invalid syntax") {
 					// an error occured and it was not from passing a word in to atoi
 					fmt.Println("Error convering input to cage card number")
 					fmt.Println(err)
 					continue
 				}
-
 				// a misread on cc means the value 0 init
 				if cc != 0 {
 					err := deactivateCC(cfg, cc, date)
 					if err != nil {
 						fmt.Println(err)
 					}
+					// don't need to visit the switch, one input is assumed to be a cc#
+					continue
 				}
 
 			case "print":
@@ -189,28 +166,6 @@ func deactivateFunction(cfg *Config) error {
 
 			case "help":
 				cmdHelp(flags)
-
-				/* removed because cards are no longer kept in a queue
-				case "pop":
-					length := len(cardsToDeactivate)
-					if length == 0 {
-						fmt.Println("No cards have been entered")
-						break
-					}
-					popped := cardsToDeactivate[length-1]
-					fmt.Printf("Popped %v\n", popped.CcID)
-					cardsToDeactivate = cardsToDeactivate[0 : length-1]
-				*/
-
-				/* removed because cards are no longer kept in a queue
-				case "process":
-					fmt.Println("Processing...")
-					err := deactivateCageCards(cfg, cardsToDeactivate)
-					if err != nil {
-						fmt.Println(err)
-					}
-					exit = true
-				*/
 
 			case "exit":
 				fmt.Println("Exiting without saving...")
@@ -230,13 +185,11 @@ func deactivateFunction(cfg *Config) error {
 	return nil
 }
 
-// yeah, just the date. Keep the 'deactivated_by' hidden
 func printCurrentDeactivationParams(date *time.Time) {
 	fmt.Println("Current settings for cards being added to deactivation queue:")
 	fmt.Printf("Date: %v\n", *date)
 }
 
-// TODO: naming things is hard. really just need the ccID huh
 func deactivateCC(cfg *Config, ccID int, date time.Time) error {
 	cc, err := cfg.db.GetCageCardByID(context.Background(), int32(ccID))
 	if err != nil && err.Error() == "sql: no rows in result set" {
@@ -300,137 +253,3 @@ func deactivateCC(cfg *Config, ccID int, date time.Time) error {
 
 	return nil
 }
-
-/* removed because no longer processing CCs via a queue
-// is it more expensive to pass an int by pointer and deref or just pass by value
-func getCCToDeactivate(cc int, date *time.Time, deactivatedBy *database.Investigator) database.DeactivateCageCardParams {
-	tdate := sql.NullTime{Valid: true, Time: *date}
-	tdeactivatedBy := uuid.NullUUID{Valid: true, UUID: deactivatedBy.ID}
-
-	tdccp := database.DeactivateCageCardParams{
-		CcID:          int32(cc),
-		DeactivatedOn: tdate,
-		DeactivatedBy: tdeactivatedBy,
-	}
-
-	return tdccp
-}
-
-// candidate for DRYing up with a "process cc" function with an activate/deactivate enum
-// behavior is just different enough to have them disentangled
-// ie activating checks to see if it's already active, deact checks to see if it isnt active
-func deactivateCageCards(cfg *Config, ctd []database.DeactivateCageCardParams) error {
-	if len(ctd) == 0 {
-		return errors.New("oops! No cards")
-	}
-	deactivationErrors := []ccError{}
-	totalDeactivated := 0
-
-	for _, cc := range ctd {
-		ccErr := checkDeactivateError(cfg, &cc)
-		// hacky way to see if a nil struct was returned, meaning no error
-		if ccErr.CCid != 0 {
-			deactivationErrors = append(deactivationErrors, ccErr)
-			continue
-		}
-
-		dcc, err := cfg.db.DeactivateCageCard(context.Background(), cc)
-		if err != nil {
-			// any other error
-			tcce := ccError{
-				CCid: int(dcc.CcID),
-				Err:  err.Error(),
-			}
-			deactivationErrors = append(deactivationErrors, tcce)
-			continue
-		}
-
-		if verbose {
-			fmt.Println(dcc)
-		}
-
-		totalDeactivated++
-
-	}
-
-	fmt.Printf("%v cards deactivated\n", totalDeactivated)
-	if len(deactivationErrors) > 0 {
-		fmt.Println("Errors deactivating these cage cards:")
-		for _, cce := range deactivationErrors {
-			fmt.Printf("%v -- %s\n", cce.CCid, cce.Err)
-		}
-	}
-	return nil
-}
-
-
-// TODO: maybe add a check for if deactivation date is after today too
-// like can only deactivate today or past, not future, to prevent errors of course
-// no "this card will have had been deactivated"
-// WORKING: seeing why stopping cards isn't working
-// I FORGOT TO FINISH THE CASES LE MOO ALSO ADD A DATE SETTER AND CHECK IF ITS IN THE FUTURE
-func checkDeactivateError(cfg *Config, cc *database.DeactivateCageCardParams) ccError {
-	// check if already active
-	td, err := cfg.db.GetActivationDate(context.Background(), cc.CcID)
-	if err != nil && err.Error() == "sql: no rows in result set" {
-		// cc not added to db or not found
-		tcce := ccError{
-			CCid: int(cc.CcID),
-			Err:  "CC not added to database",
-		}
-
-		return tcce
-	}
-
-	if !td.Valid {
-		tcce := ccError{
-			CCid: int(cc.CcID),
-			Err:  "CC is not currently active",
-		}
-		return tcce
-	}
-
-	// check if deactivation date is before activation date
-	if cc.DeactivatedOn.Time.Before(td.Time) {
-		tcce := ccError{
-			CCid: int(cc.CcID),
-			Err:  "Deactivation date is before activation date",
-		}
-		return tcce
-	}
-
-	if err != nil {
-		// any other error
-		tcce := ccError{
-			CCid: int(cc.CcID),
-			Err:  err.Error(),
-		}
-		return tcce
-	}
-
-	// check if previously deactivated
-	dd, err := cfg.db.GetDeactivationDate(context.Background(), cc.CcID)
-	// dont need to check if not in db
-	if dd.Valid {
-		// card was previously deactivated
-		errmsg := fmt.Sprintf("CC is already deactivated -- %s", dd.Time)
-		tcce := ccError{
-			CCid: int(cc.CcID),
-			Err:  errmsg,
-		}
-		return tcce
-	}
-
-	if err != nil {
-		// any other error
-		tcce := ccError{
-			CCid: int(cc.CcID),
-			Err:  err.Error(),
-		}
-		return tcce
-	}
-
-	// everything ok
-	return ccError{}
-}
-*/
