+++ cmd_investigator.go
@@ -4,12 +4,10 @@ import (
 	"bufio"
 	"context"
 	"database/sql"
-	"errors"
 	"fmt"
 	"os"
 	"strings"
 
-	"github.com/google/uuid"
 	"github.com/jsandberg07/clitest/internal/database"
 )
 
@@ -26,8 +24,6 @@ func getEditInvestigatorCmd() Command {
 	return editInvestigatorCmd
 }
 
-// print, exit, save, help, [i]nvestigator name, [n]ickname, [p]osition, [a]ctive, [e]mail
-// TODO: a function that parses things with QUOTES and a space properly JK JUST USE UNDERSCORES
 func getEditInvestigatorFlags() map[string]Flag {
 	editInvestigatorFlags := make(map[string]Flag)
 	iFlag := Flag{
@@ -70,8 +66,6 @@ func getEditInvestigatorFlags() map[string]Flag {
 	}
 	editInvestigatorFlags[eFlag.symbol] = eFlag
 
-	// ect as needed or remove the "-"+ for longer ones
-
 	printFlag := Flag{
 		symbol:      "print",
 		description: "Prints current data for review",
@@ -108,11 +102,7 @@ func getEditInvestigatorFlags() map[string]Flag {
 
 }
 
-// ask for a name, then pass in flags for everything
-// then print old to new
-// look into removing the args thing, might have to stay
 func editInvestigatorFunction(cfg *Config) error {
-	// permission check
 	err := checkPermission(cfg.loggedInPosition, PermissionStaff)
 	if err != nil {
 		return err
@@ -132,10 +122,8 @@ func editInvestigatorFunction(cfg *Config) error {
 		os.Exit(1)
 	}
 
-	// get flags
 	flags := getEditInvestigatorFlags()
 
-	// set defaults
 	exit := false
 	reviewed := Reviewed{
 		Printed:     false,
@@ -150,14 +138,12 @@ func editInvestigatorFunction(cfg *Config) error {
 		Active:   investigator.Active,
 	}
 
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
 
 	fmt.Printf("Making changes to %s\n", investigator.IName)
 	fmt.Println("If entering a value that uses spaces (like between a first and last name) use an underscore")
 	fmt.Println("It will be changed to a space after")
 
-	// da loop
 	for {
 		fmt.Print("> ")
 		text, err := reader.ReadString('\n')
@@ -172,29 +158,27 @@ func editInvestigatorFunction(cfg *Config) error {
 			continue
 		}
 
-		// do weird behavior here
 		if reviewed.ChangesMade {
 			reviewed.Printed = false
 		}
 
-		// but normal loop now
 		args, err := parseArguments(flags, inputs)
 		if err != nil {
 			fmt.Println(err)
 			continue
 		}
 
-		// print, exit, save, help, [i]nvestigator name, [n]ickname, [p]osition, [a]ctive, [e]mail
 		for _, arg := range args {
 			switch arg.flag {
 			case "print":
-				// need some generic type for printing
 				printEditInvestigator(&uiParam, &position)
 				reviewed.ChangesMade = false
 				reviewed.Printed = true
+
 			case "exit":
 				fmt.Println("Exiting without saving...")
 				exit = true
+
 			case "save":
 				if reviewed.Printed {
 					fmt.Println("Saving...")
@@ -210,8 +194,10 @@ func editInvestigatorFunction(cfg *Config) error {
 				}
 				fmt.Println("Investigator has been added. They will be asked to update their password on first login")
 				exit = true
+
 			case "help":
 				cmdHelp(flags)
+
 			case "-i":
 				err := checkIfInvestigatorNameUnique(cfg, arg.value)
 				if err != nil {
@@ -219,6 +205,7 @@ func editInvestigatorFunction(cfg *Config) error {
 				}
 				uiParam.IName = arg.value
 				reviewed.ChangesMade = true
+
 			case "-n":
 				if arg.value == "delete" {
 					uiParam.Nickname = sql.NullString{Valid: false}
@@ -226,13 +213,16 @@ func editInvestigatorFunction(cfg *Config) error {
 					uiParam.Nickname = sql.NullString{Valid: true, String: arg.value}
 				}
 				reviewed.ChangesMade = true
+
 			case "-p":
 				position, err = getPositionByFlag(cfg, arg.value)
 				if err != nil {
-					return err
+					fmt.Println(err)
+					continue
 				}
 				uiParam.Position = position.ID
 				reviewed.ChangesMade = true
+
 			case "-a":
 				uiParam.Active = !uiParam.Active
 				if uiParam.Active {
@@ -241,6 +231,7 @@ func editInvestigatorFunction(cfg *Config) error {
 					fmt.Println("Invetstigator is flagged as inactive")
 				}
 				reviewed.ChangesMade = true
+
 			case "-e":
 				if arg.value == "delete" {
 					uiParam.Email = sql.NullString{Valid: false}
@@ -248,17 +239,15 @@ func editInvestigatorFunction(cfg *Config) error {
 					uiParam.Email = sql.NullString{Valid: true, String: arg.value}
 				}
 				reviewed.ChangesMade = true
+
 			default:
 				fmt.Printf("%s%s\n", DefaultFlagMsg, arg.flag)
 			}
 		}
-
 		if exit {
 			break
 		}
-
 	}
-
 	return nil
 }
 
@@ -274,53 +263,6 @@ func printEditInvestigator(ui *database.UpdateInvestigatorParams, p *database.Po
 	if ui.Nickname != nullfield {
 		fmt.Printf("Nickname: %s\n", ui.Nickname.String)
 	}
-}
-
-func getPositionByFlag(cfg *Config, title string) (database.Position, error) {
-	position, err := cfg.db.GetPositionByTitle(context.Background(), title)
-	if err != nil && err.Error() != "sql: no rows in result set" {
-		fmt.Println("Error getting position from DB")
-		return database.Position{ID: uuid.Nil}, err
-	}
-	if err.Error() == "sql: no rows in result set" {
-		fmt.Println("No position by that title found")
-		return database.Position{ID: uuid.Nil}, err
-	}
-
-	return position, nil
-
-}
-
-func checkIfInvestigatorNameUnique(cfg *Config, name string) error {
-	investigators, err := cfg.db.GetInvestigatorByName(context.Background(), name)
-	if err != nil && err.Error() != "sql: no rows in result set" {
-		fmt.Println("Error getting name from DB")
-		return err
-	}
-	if len(investigators) != 0 {
-		fmt.Println("Investigator name is not unique. Please consider adding a nickname to both investigators.")
-	}
-	return nil
-}
-
-func getInvestigatorStruct(cfg *Config, input string) (database.Investigator, error) {
-	investigators, err := cfg.db.GetInvestigatorByName(context.Background(), input)
-	if err != nil && err.Error() == "sql: no rows in result set" {
-		return database.Investigator{}, errors.New("investigator not found. Please try again")
-	}
-	// TODO: does returning many even throw the "no rows in result set" error?
-	if len(investigators) == 0 {
-		return database.Investigator{}, errors.New("investigator not found. Please try again")
-	}
-	if len(investigators) > 1 {
-		return database.Investigator{}, errors.New("vague investigator name. Please try again")
-	}
-	if err != nil {
-		// any other error
-		return database.Investigator{}, err
-	}
-
-	return investigators[0], nil
 }
 
 func getAddInvestigatorCmd() Command {
@@ -362,17 +304,9 @@ func getAddInvestigatorFlags() map[string]Flag {
 	}
 	addInvestigatorFlags[helpFlag.symbol] = helpFlag
 
-	// ect as needed or remove the "-"+ for longer ones
-
 	return addInvestigatorFlags
-
 }
 
-// prompt step by step cause some of the fields could be null and whatever
-// then review and you can change them with flags after
-// and always allow you to exit early or whatever
-// name, position, email, nickname, assume active
-// look into removing the args thing, might have to stay
 func addInvestigatorFunction(cfg *Config) error {
 	err := checkPermission(cfg.loggedInPosition, PermissionStaff)
 	if err != nil {
@@ -405,22 +339,17 @@ func addInvestigatorFunction(cfg *Config) error {
 		return err
 	}
 
-	// get flags
 	flags := getAddInvestigatorFlags()
 
-	// set defaults
 	exit := false
 
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
 	fmt.Println("New investigator will be added with the following info:")
 	printNewInvestigator(&ciParams, &position)
 
-	// TODO: add the ability to use flags to edit values after the fact
 	fmt.Println("Enter 'save' to add the investigator to the database")
 	fmt.Println("Or 'exit' to exit without saving")
 
-	// da loop
 	for {
 		fmt.Print("> ")
 		text, err := reader.ReadString('\n')
@@ -435,9 +364,6 @@ func addInvestigatorFunction(cfg *Config) error {
 			continue
 		}
 
-		// do weird behavior here
-
-		// but normal loop now
 		args, err := parseArguments(flags, inputs)
 		if err != nil {
 			fmt.Println(err)
@@ -467,24 +393,20 @@ func addInvestigatorFunction(cfg *Config) error {
 				fmt.Printf("%s%s\n", DefaultFlagMsg, arg.flag)
 			}
 		}
-
 		if exit {
 			break
 		}
-
 	}
-
 	return nil
 }
 
+// separate because it allows empty inputs
 func getNewInvestigatorExtraInfo(ciParams *database.CreateInvestigatorParams) error {
 	fmt.Println("The next info isn't required. Skip by entering nothing.")
 	fields := []string{"a nickname", "an email"}
 	inputs := make([]string, len(fields))
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
 
-	// da loop
 	for i, field := range fields {
 		fmt.Printf("Please enter %s\n", field)
 		fmt.Print("> ")
@@ -525,116 +447,3 @@ func printNewInvestigator(ci *database.CreateInvestigatorParams, p *database.Pos
 		fmt.Printf("Nickname: %s\n", ci.Nickname.String)
 	}
 }
-
-/* removed in refactor
-
-func getNewInvestigatorName(cfg *Config) (string, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the name of the investigator, or exit to cancel")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again.")
-			continue
-		}
-		if input == "exit" || input == "cancel" {
-			return "", nil
-		}
-
-		err = checkIfInvestigatorNameUnique(cfg, input)
-		if err != nil {
-			return "", err
-		}
-
-		return input, nil
-
-
-	}
-}
-*/
-
-/* removed from refactor
-
-
-func getInvestigatorByName(cfg *Config) (database.Investigator, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the name or nickname of the investigator you would like to edit, or exit to cancel")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again.")
-			continue
-		}
-		if input == "exit" || input == "cancel" {
-			return database.Investigator{ID: uuid.Nil}, nil
-		}
-
-		investigators, err := cfg.db.GetInvestigatorByName(context.Background(), input)
-		if err != nil && err.Error() != "sql: no rows in result set" {
-			// error that isnt related to no rows returned
-			fmt.Println("Error checking db for investigator")
-			return database.Investigator{ID: uuid.Nil}, err
-		}
-		if len(investigators) == 0 {
-			fmt.Println("No investigator by that name or nickname found. Please try again.")
-			continue
-		}
-		if len(investigators) > 1 {
-			fmt.Println("Vague investigator name. Consider entering a nickname instead.")
-			continue
-		}
-		return investigators[0], nil
-	}
-}
-
-
-
-*/
-
-/* removed by refactor???
-func getNewInvestigatorPosition(cfg *Config) (database.Position, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the position of the new investigator, or exit to cancel")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again.")
-			continue
-		}
-		if input == "exit" || input == "cancel" {
-			return database.Position{ID: uuid.Nil}, nil
-		}
-
-		position, err := cfg.db.GetPositionByTitle(context.Background(), input)
-		if err != nil && err.Error() == "sql: no rows in result set" {
-			fmt.Println("Position by that title not found. Please try again")
-			continue
-		}
-		if err != nil {
-			fmt.Println("Error getting position from database")
-			return database.Position{ID: uuid.Nil}, err
-		}
-
-		return position, nil
-
-	}
-}
-*/
