+++ cmd_protocol.go
@@ -3,7 +3,6 @@ package main
 import (
 	"bufio"
 	"context"
-	"errors"
 	"fmt"
 	"os"
 	"strconv"
@@ -26,8 +25,6 @@ func getAddProtocolCmd() Command {
 	return addProtocolCmd
 }
 
-// TODO: add editing it before saving it, goes hand in hand with editing but that can wait
-// just save and exit in the mean time
 func getAddProtocolFlags() map[string]Flag {
 	addProtocolFlags := make(map[string]Flag)
 	saveFlag := Flag{
@@ -58,23 +55,17 @@ func getAddProtocolFlags() map[string]Flag {
 
 }
 
-// look into removing the args thing, might have to stay
 func addProtocolFunction(cfg *Config) error {
-	// permission check
 	err := checkPermission(cfg.loggedInPosition, PermissionProtocol)
 	if err != nil {
 		return err
 	}
-	// get flags
 	flags := getAddProtocolFlags()
 
-	// set defaults
 	exit := false
 
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
 
-	// TODO: not all of these even have errs that can be returned, remove as needed
 	pi, err := getStructPrompt(cfg, "Enter the name of the PI overseeing the protocol", getPIStruct)
 	if err != nil && err.Error() != CancelError {
 		return err
@@ -93,7 +84,7 @@ func addProtocolFunction(cfg *Config) error {
 		return nil
 	}
 
-	number, err := getStringPrompt(cfg, "Enter number of new protocol", checkProtocolUniqueFunc)
+	number, err := getStringPrompt(cfg, "Enter number of new protocol", checkProtocolUnique)
 	if err != nil && err.Error() != CancelError {
 		return err
 	}
@@ -126,8 +117,6 @@ func addProtocolFunction(cfg *Config) error {
 		ExpirationDate:      expiration,
 	}
 
-	// save and exit for now
-	// da loop
 	fmt.Println("Current info:")
 	printAddProtocol(cpParam, pi)
 	fmt.Println("Enter 'save' to save, 'exit' to exit without saving")
@@ -145,9 +134,6 @@ func addProtocolFunction(cfg *Config) error {
 			continue
 		}
 
-		// do weird behavior here
-
-		// but normal loop now
 		args, err := parseArguments(flags, inputs)
 		if err != nil {
 			fmt.Println(err)
@@ -186,16 +172,6 @@ func addProtocolFunction(cfg *Config) error {
 	return nil
 }
 
-// TODO: CURRENTLY 0 restriction on what positions can be a PI
-// get everything else going first
-// requires a DB change, add a field for like name_on_protocol or is_pi_protocol for position
-// just another prompt and tag that should be easy, then test data included
-
-// TODO: enter previous protocol FIRST, copies a lot of work, updates cards
-// increasing todo list buddy
-
-// TODO: add check if protocol by that title already in the DB (number more important really)
-
 func getNewProtocolExpiration() (time.Time, error) {
 	reader := bufio.NewReader(os.Stdin)
 	fmt.Println("Enter expiration date of the protocol")
@@ -232,46 +208,6 @@ func printAddProtocol(cp database.CreateProtocolParams, pi database.Investigator
 	fmt.Printf("Expiration Date: %v\n", cp.ExpirationDate)
 }
 
-func getPIStruct(cfg *Config, input string) (database.Investigator, error) {
-	investigators, err := cfg.db.GetInvestigatorByName(context.Background(), input)
-	if err != nil && err.Error() == "sql: no rows in result set" {
-		return database.Investigator{}, errors.New("investigator not found. Please try again")
-	}
-	// TODO: does returning many even throw the "no rows in result set" error?
-	if len(investigators) == 0 {
-		return database.Investigator{}, errors.New("investigator not found. Please try again")
-	}
-	if len(investigators) > 1 {
-		return database.Investigator{}, errors.New("vague investigator name. Please try again")
-	}
-	if err != nil {
-		// any other error
-		return database.Investigator{}, err
-	}
-
-	// TODO: add checking permissions to see if this person can oversee a protocol
-
-	return investigators[0], err
-}
-
-func checkProtocolUniqueFunc(cfg *Config, input string) error {
-	protocol, err := cfg.db.GetProtocolByNumber(context.Background(), input)
-	if err != nil && err.Error() == "sql: no rows in result set" {
-		// input is unique
-		return nil
-	}
-	if err != nil {
-		// any other error
-		fmt.Println("Error checking db for protocol.")
-		return err
-	}
-
-	// not unique
-	fmt.Printf("Protocol with same number: %s\n", protocol.Title)
-	return errors.New("a protocol with that number already exists. Please try again")
-
-}
-
 func getEditProtocolCmd() Command {
 	editProtocolFlags := make(map[string]Flag)
 	editProtocolCmd := Command{
@@ -285,8 +221,6 @@ func getEditProtocolCmd() Command {
 	return editProtocolCmd
 }
 
-// [t]itle, [P]I, [N]umber, [A]llocated, [B]alance, [E]xpiration
-// save, exit, help
 func getEditProtocolFlags() map[string]Flag {
 	editProtocolFlags := make(map[string]Flag)
 	tFlag := Flag{
@@ -329,8 +263,6 @@ func getEditProtocolFlags() map[string]Flag {
 	}
 	editProtocolFlags[eFlag.symbol] = eFlag
 
-	// ect as needed or remove the "-"+ for longer ones
-
 	exitFlag := Flag{
 		symbol:      "exit",
 		description: "Exits without saving changes",
@@ -367,9 +299,7 @@ func getEditProtocolFlags() map[string]Flag {
 
 }
 
-// look into removing the args thing, might have to stay
 func editProtocolFunction(cfg *Config) error {
-	// permission check
 	err := checkPermission(cfg.loggedInPosition, PermissionProtocol)
 	if err != nil {
 		return err
@@ -398,23 +328,18 @@ func editProtocolFunction(cfg *Config) error {
 		return err
 	}
 
-	// get flags
 	flags := getEditProtocolFlags()
 
-	// set defaults
 	exit := false
 	reviewed := Reviewed{
 		Printed:     false,
 		ChangesMade: false,
 	}
 
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
 	fmt.Println("Use flags to change protocol parameters. Enter 'help' to see all available flags")
 	fmt.Println("When entering values with a space, replace it with an underscore")
 
-	// da loop
-	// [t]itle, [P]I, [N]umber, [A]llocated, [B]alance, [E]xpiration
 	for {
 		fmt.Print("> ")
 		text, err := reader.ReadString('\n')
@@ -429,12 +354,10 @@ func editProtocolFunction(cfg *Config) error {
 			continue
 		}
 
-		// do weird behavior here
 		if reviewed.ChangesMade {
 			reviewed.Printed = false
 		}
 
-		// but normal loop now
 		args, err := parseArguments(flags, inputs)
 		if err != nil {
 			fmt.Println(err)
@@ -446,26 +369,27 @@ func editProtocolFunction(cfg *Config) error {
 			case "-t":
 				upParams.Title = arg.value
 				reviewed.ChangesMade = true
+
 			case "-p":
-				// updating the pi outside of the block, but will it mess with the error check? glass coffin
 				pi, err = getInvestigatorByFlag(cfg, arg.value)
 				if err != nil {
 					return err
 				}
 				upParams.PrimaryInvestigator = pi.ID
 				reviewed.ChangesMade = true
+
 			case "-n":
 				if arg.value == protocol.PNumber {
-					// same as the OG, so it'll throw an error if trying to change back
 					upParams.PNumber = arg.value
 					break
 				}
-				number, err := checkIfProtocolNumberUnique(cfg, arg.value)
+				number, err := getUniqueProtocolFromFlag(cfg, arg.value)
 				if err != nil {
 					return err
 				}
 				upParams.PNumber = number
 				reviewed.ChangesMade = true
+
 			case "-a":
 				allocated, err := strconv.Atoi(arg.value)
 				if err != nil {
@@ -474,6 +398,7 @@ func editProtocolFunction(cfg *Config) error {
 				}
 				upParams.Allocated = int32(allocated)
 				reviewed.ChangesMade = true
+
 			case "-b":
 				balance, err := strconv.Atoi(arg.value)
 				if err != nil {
@@ -482,6 +407,7 @@ func editProtocolFunction(cfg *Config) error {
 				}
 				upParams.Balance = int32(balance)
 				reviewed.ChangesMade = true
+
 			case "-e":
 				date, err := parseDate(arg.value)
 				if err != nil {
@@ -494,15 +420,19 @@ func editProtocolFunction(cfg *Config) error {
 				}
 				upParams.ExpirationDate = date
 				reviewed.ChangesMade = true
+
 			case "help":
 				cmdHelp(flags)
+
 			case "print":
 				printEditProtocol(&upParams, &pi)
 				reviewed.ChangesMade = false
 				reviewed.Printed = true
+
 			case "exit":
 				fmt.Println("Exiting without saving...")
 				exit = true
+
 			case "save":
 				fmt.Println("Saving...")
 				err := cfg.db.UpdateProtocol(context.Background(), upParams)
@@ -530,68 +460,12 @@ func editProtocolFunction(cfg *Config) error {
 	return nil
 }
 
-// [t]itle, [P]I, [N]umber, [A]llocated, [B]alance, [E]xpiration, is [A]ctive
-func getInvestigatorByFlag(cfg *Config, i string) (database.Investigator, error) {
-	investigators, err := cfg.db.GetInvestigatorByName(context.Background(), i)
-	if err != nil && err.Error() != "sql: no rows in result set" {
-		// any other error
-		fmt.Println("Error getting investigators from DB")
-		return database.Investigator{}, err
-	}
-	if err != nil && err.Error() == "sql: no rows in result set" {
-		fmt.Println("Investigator not found. Please try again")
-		return database.Investigator{}, nil
-	}
-	if len(investigators) > 1 {
-		fmt.Println("Vague investigator name. Please try again")
-		return database.Investigator{}, nil
-	}
-	if len(investigators) == 0 {
-		fmt.Println("Investigator not found. Please try again")
-		return database.Investigator{}, nil
-	}
-	return investigators[0], nil
-}
-
-func checkIfProtocolNumberUnique(cfg *Config, n string) (string, error) {
-	_, err := cfg.db.GetProtocolByNumber(context.Background(), n)
-	if err != nil && err.Error() != "sql: no rows in result set" {
-		// any other error
-		fmt.Println("Error getting protocols from DB")
-		return "", err
-	}
-	if err == nil {
-		// protocol found
-		fmt.Println("Protocol with that number already exists. Please try again")
-		return "", err
-	}
-
-	// if nothing found, then unique and ok
-	return n, nil
-
-}
-
-// look into having it accept one of two generic types as uses same values to print as createProtParam
 func printEditProtocol(up *database.UpdateProtocolParams, pi *database.Investigator) {
 	fmt.Printf("PI: %s\n", pi.IName)
 	fmt.Printf("Number: %s\n", up.PNumber)
 	fmt.Printf("Title: %s\n", up.Title)
 	fmt.Printf("Allocated: %v\n", up.Allocated)
 	fmt.Printf("Expiration Date: %v\n", up.ExpirationDate)
-}
-
-func checkProtocolExists(cfg *Config, input string) (database.Protocol, error) {
-	protocol, err := cfg.db.GetProtocolByNumber(context.Background(), input)
-	if err != nil && err.Error() == "sql: no rows in result set" {
-		// not found
-		return database.Protocol{}, errors.New("no protocol with that number found. Please try again")
-	}
-	if err != nil {
-		// any other error
-		return database.Protocol{}, err
-	}
-
-	return protocol, nil
 }
 
 func addBalanceToProtocol(cfg *Config, input int, cc *database.CageCard) error {
@@ -604,7 +478,6 @@ func addBalanceToProtocol(cfg *Config, input int, cc *database.CageCard) error {
 		return err
 	}
 
-	// TODO: is there a way to get psql to return a bool for balance > allocated?
 	protocol, err := cfg.db.GetProtocolByID(context.Background(), cc.ProtocolID)
 	if err != nil {
 		return err
@@ -616,163 +489,3 @@ func addBalanceToProtocol(cfg *Config, input int, cc *database.CageCard) error {
 
 	return nil
 }
-
-/* removed via refactor
-func getProtocolByNumber(cfg *Config) (database.Protocol, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the number of the protocol you'd like to edit, or exit to cancel")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again.")
-			continue
-		}
-		if input == "exit" || input == "cancel" {
-			return database.Protocol{ID: uuid.Nil}, nil
-		}
-
-		protocol, err := cfg.db.GetProtocolByNumber(context.Background(), input)
-		if err != nil && err.Error() != "sql: no rows in result set" {
-			// error that isnt related to no rows returned
-			fmt.Println("Error checking DB for protocol")
-			return database.Protocol{ID: uuid.Nil}, err
-		}
-		if err != nil && err.Error() == "sql: no rows in result set" {
-			fmt.Println("No protocol with that number found. Please try again.")
-			continue
-		}
-
-		return protocol, nil
-	}
-}
-*/
-
-/* removed in refactor
-func getNewProtocolPI(cfg *Config) (database.Investigator, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the name of the PI overseeing the protocol")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again")
-			continue
-		}
-
-		investigators, err := cfg.db.GetInvestigatorByName(context.Background(), input)
-		if err != nil && err.Error() != "sql: no rows in result set" {
-			// error that isnt related to no rows returned
-			fmt.Println("Error checking db for investigator")
-			return database.Investigator{ID: uuid.Nil}, err
-		}
-		if len(investigators) == 0 {
-			fmt.Println("No investigator by that name or nickname found. Please try again.")
-			continue
-		}
-		if len(investigators) > 1 {
-			fmt.Println("Vague investigator name. Consider entering a nickname instead.")
-			continue
-		}
-		return investigators[0], nil
-
-	}
-}
-*/
-
-/* removed in refactor
-func getNewProtocolTitle() (string, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the title of the protocol")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again")
-			continue
-		}
-
-		return input, nil
-	}
-
-}
-*/
-
-/* removed via refactor
-func getNewProtocolNumber(cfg *Config) (string, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the protocol number")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again")
-			continue
-		}
-
-		// TODO: this is actually protocol by number, not ID, needs fixing
-		protocol, err := cfg.db.GetProtocolByNumber(context.Background(), input)
-		if err != nil && err.Error() != "sql: no rows in result set" {
-			// any other error than no rows
-			fmt.Println("Error getting protocol from DB")
-			return "", err
-		}
-		if err != nil && err.Error() == "sql: no rows in result set" {
-			// no rows, which is ideal
-			return input, nil
-		}
-		if err == nil {
-			fmt.Println("A protocol with that number already exists. Please try again.")
-			fmt.Printf("Protocol with same number: %s\n", protocol.Title)
-		}
-	}
-}
-*/
-
-/* removed via refactor
-func getNewProtocolAlocated() (int, error) {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Println("Enter the numbers of animals allocated to the protocol")
-	for {
-		fmt.Print("> ")
-		text, err := reader.ReadString('\n')
-		if err != nil {
-			fmt.Printf("Error reading input string: %s", err)
-			os.Exit(1)
-		}
-		input := strings.TrimSpace(text)
-		if input == "" {
-			fmt.Println("No input found. Please try again")
-			continue
-		}
-
-		allocated, err := strconv.Atoi(input)
-		if err != nil {
-			fmt.Println("Error getting number. Be sure to enter an integer.")
-			continue
-		}
-
-		return allocated, nil
-	}
-}
-*/
