+++ cmd_queries_cc.go
@@ -26,40 +26,6 @@ func getCCQueriesCmd() Command {
 
 	return CCQueriesCmd
 }
-
-// 5 functions
-// PI (skip)
-// protocol
-// investigator
-// pi + investigator (skip)
-// protocol + investigator (skip this as well for now)
-
-// then set start and end dates
-
-// bonus ones:
-// all active
-// ALL cage cards
-
-// write sql for each type of query
-// write the go for setting the tags
-// write the normalizer functions (have to because psql will probably reset them)
-// the data is all the same but the name is different
-// so return that, string that, export that
-// pass a function first class, normalize response, return, stringify, export
-// why do i have a dozen "true activate" "real activate" "honest activate" -_-
-// i have literally spend tons of time looking through each step BEFORE this
-// gotta plan the whole thing next time lol
-
-// FACT: i didn't make it so i can get them by PI because it created vague returns in postgres so skip it for now lmao
-// TODO: get the PI as well in the export
-// TODO: get the note in the export
-// TODO: reset the protoocl or investigator with like x or whatever (hope nobody is named X lmao)
-// TODO: make sure the dates get the correct data, since they're set at like midnigt or something
-
-// so we need to print the order number too, last thing for today i guess and then make a plan for tomorrow
-//
-
-// [s]tart date, [e]nd date, [pr]otocol, [in]vestigator, print, active, all, help, exit, query
 func getCCQueriesFlags() map[string]Flag {
 	ccQueriesFlags := make(map[string]Flag)
 
@@ -102,8 +68,6 @@ func getCCQueriesFlags() map[string]Flag {
 		printOrder:  6,
 	}
 	ccQueriesFlags[orFlag.symbol] = orFlag
-
-	// ect as needed or remove the "-"+ for longer ones
 
 	printFlag := Flag{
 		symbol:      "print",
@@ -157,37 +121,26 @@ func getCCQueriesFlags() map[string]Flag {
 
 }
 
-// you were working on adding tags, writing a few functions for getting the data
-// and a struct normalizer because i'll do that instead of changing the return values
-
-// look into removing the args thing, might have to stay
-// TODO: create a struct with an enum that remembers the query type, nulls the current values when you set a new one,
-// is easy to pass, run a switch on it
 func CCQueriesFunction(cfg *Config) error {
-	// permission check
 	err := checkPermission(cfg.loggedInPosition, PermissionRunQueries)
 	if err != nil {
 		return err
 	}
-	// get flags
+
 	flags := getCCQueriesFlags()
 
-	// set defaults
 	exit := false
 	investigator := database.Investigator{}
 	protocol := database.Protocol{}
 	order := database.Order{}
 
-	// might be a problem since data is stored at midnight, might have to do some rounding
 	startDate := normalizeDate(time.Now())
 	endDate := normalizeDate(time.Now())
 
-	// the reader
 	reader := bufio.NewReader(os.Stdin)
+
 	fmt.Println("Enter a protocol or investigator, date range to get all cards active during that time frame")
 	fmt.Println("Or 'active' to get all currently active cage cards")
-
-	// da loop
 	for {
 		fmt.Print("> ")
 		text, err := reader.ReadString('\n')
@@ -202,17 +155,12 @@ func CCQueriesFunction(cfg *Config) error {
 			continue
 		}
 
-		// do weird behavior here
-
-		// but normal loop now
 		args, err := parseArguments(flags, inputs)
 		if err != nil {
 			fmt.Println(err)
 			continue
 		}
 
-		// [s]tart date, [e]nd date, [pr]otocol, [in]vestigator, print, active, all, help, exit, query
-		// add an [or] for order
 		for _, arg := range args {
 			switch arg.flag {
 
@@ -298,7 +246,6 @@ func CCQueriesFunction(cfg *Config) error {
 				}
 				exit = true
 
-				// TODO: replace this with a switch, struct that stores the values (and can be reset), and a function call
 			case "query":
 				nilInvestigator := database.Investigator{}
 				nilProtocol := database.Protocol{}
@@ -359,16 +306,6 @@ func CCQueriesFunction(cfg *Config) error {
 	return nil
 }
 
-// keep it simple
-// different function for option
-// pass in the params
-// run the query
-// normalize the result
-// then pass it to export
-// so write a function for each query
-// normalize for each struct
-// one export
-
 func CCQueryActive(cfg *Config) error {
 	ccs, err := cfg.db.GetCageCardsActive(context.Background())
 	if err != nil {
@@ -415,12 +352,6 @@ func CCQueryAll(cfg *Config) error {
 }
 
 func CCQueryDateRange(cfg *Config, start, end time.Time) error {
-	/*
-		gcdrParam := database.GetCardsDateRangeParams{
-			ActivatedOn:   sql.NullTime{Valid: true, Time: start},
-			DeactivatedOn: sql.NullTime{Valid: true, Time: end},
-		}
-	*/
 	gcdrParam := database.GetCardsDateRangeParams{
 		Overlaps:   sql.NullTime{Valid: true, Time: start},
 		Overlaps_2: sql.NullTime{Valid: true, Time: end},
@@ -475,7 +406,6 @@ func CCQueryInvestigator(cfg *Config, start, end time.Time, inv *database.Invest
 
 }
 
-// TODO: these are all similar. When you have a switch, these will get condensed into 1 function surely
 func CCQueryOrder(cfg *Config, o *database.Order) error {
 	ccs, err := cfg.db.GetCageCardsOrder(context.Background(), uuid.NullUUID{Valid: true, UUID: o.ID})
 	if err != nil {
@@ -487,7 +417,6 @@ func CCQueryOrder(cfg *Config, o *database.Order) error {
 		return nil
 	}
 
-	// TODO: adding a column to this is a bad experience. Modifying like 5 structs via sql functions.
 	exp := NormalizeCCExport(ccs)
 
 	count, err := exportData(exp)
@@ -497,21 +426,6 @@ func CCQueryOrder(cfg *Config, o *database.Order) error {
 
 	fmt.Printf("// Exported lines: %v\n", count)
 	return nil
-}
-
-// Output rows are consistent between types, but sqlc generates new struct for each query.
-// Changes them into a format that can be turned into a string to be exported. No clue what happens if the structs aren't identical!
-// Don't return error, just check if value is 0 after
-func NormalizeCCExport[T database.GetCageCardsInvestigatorRow | database.GetCageCardsAllRow | database.GetCageCardsActiveRow | database.GetCardsDateRangeRow | database.GetCageCardsProtocolRow | database.GetCageCardsOrderRow](ccs []T) []CageCardExport {
-	if len(ccs) == 0 {
-		return []CageCardExport{}
-	}
-	output := make([]CageCardExport, len(ccs))
-	for i, cc := range ccs {
-		ts := CageCardExport(cc)
-		output[i] = ts
-	}
-	return output
 }
 
 func CCQueryProtocol(cfg *Config, start, end time.Time, pro *database.Protocol) error {
@@ -541,6 +455,25 @@ func CCQueryProtocol(cfg *Config, start, end time.Time, pro *database.Protocol) 
 	fmt.Printf("// Exported lines: %v\n", count)
 	return nil
 
+}
+
+// Output rows are consistent between types, but sqlc generates new struct for each query.
+// Changes them into a format that can be turned into a string to be exported.
+func NormalizeCCExport[T database.GetCageCardsInvestigatorRow |
+	database.GetCageCardsAllRow |
+	database.GetCageCardsActiveRow |
+	database.GetCardsDateRangeRow |
+	database.GetCageCardsProtocolRow |
+	database.GetCageCardsOrderRow](ccs []T) []CageCardExport {
+	if len(ccs) == 0 {
+		return []CageCardExport{}
+	}
+	output := make([]CageCardExport, len(ccs))
+	for i, cc := range ccs {
+		ts := CageCardExport(cc)
+		output[i] = ts
+	}
+	return output
 }
 
 func exportData(cages []CageCardExport) (int, error) {
@@ -582,7 +515,6 @@ func exportData(cages []CageCardExport) (int, error) {
 	return count, nil
 }
 
-// TODO: format the dates so theyre just like a day and not a stinkin millisecond
 func stringifyCage(c *CageCardExport) []string {
 
 	output := make([]string, 7)
@@ -610,25 +542,21 @@ func stringifyCage(c *CageCardExport) []string {
 
 }
 
-// make a constant or make it changable to have the file name be consistent / alterable
 func getExportFileName() string {
 	uuid := uuid.New().String()
 	return "exports/" + uuid[0:8] + ".csv"
 }
 
 func createExportDirectory() error {
-	// err := os.Mkdir("exports", os.ModePerm)
 	err := os.Mkdir("exports", 0750)
 	if err != nil && err.Error() == "mkdir exports: file exists" {
 		// it already exists, just skip
 		return nil
-
 	}
 	if err != nil {
 		// any other error
 		fmt.Println("Error creating directory")
 		return err
 	}
-
 	return nil
 }
