+++ wishlist.md
@@ -1,0 +1,15 @@
+Things I would like to add:
+* deleting reminders after x days (and an option for it)
+* implementing the single current setting that can be changed (anybody can activate anybodys cards or add orders)
+* better instruction printing
+* explicit instructions to use "_" instead of spaces
+* more queries for all types investigators, strains, positions even 
+* add notes to CC queries 
+* better names for export files
+* spinning up a temp db for testing (https://michael.stapelberg.ch/posts/2024-11-19-testing-with-go-and-postgresql-ephemeral-dbs/)
+* PI restriction / permission
+* Printing time to just be a date (as theyre always ad midnight) 
+* Printing things when new params are set 
+* * (setFunc(type string, s interface{}) {print("x is set - values)})
+* more test data
+*  server version (yay json)
